filedelete: fix memory leaks because of attached processes
