+++ src/plugins/filedelete/filedelete.cpp
@@ -989,14 +989,8 @@ static start_readfile_t start_readfile(drakvuf_t drakvuf, drakvuf_trap_info_t* i
 
 static event_response_t createfile_ret_cb(drakvuf_t drakvuf, drakvuf_trap_info_t* info)
 {
-    addr_t attached_proc = drakvuf_get_current_attached_process(drakvuf, info);
-    proc_data_t attached_proc_data;
-    if (!attached_proc ||
-        !drakvuf_get_process_data(drakvuf, attached_proc,
-                                  &attached_proc_data) ||
-        !drakvuf_get_current_thread_id(drakvuf, info,
-                                       &attached_proc_data.tid) ||
-        !attached_proc_data.pid || !attached_proc_data.tid)
+    attached_proc_data_t proc(drakvuf, info);
+    if (!proc.pid)
     {
         PRINT_DEBUG("[FILEDELETE2] [PID:%d] [TID:%d] Error: Failed to get "
                     "attached process\n",
@@ -1006,7 +1000,7 @@ static event_response_t createfile_ret_cb(drakvuf_t drakvuf, drakvuf_trap_info_t
 
     auto w = (struct createfile_ret_info*)info->trap->data;
 
-    if (attached_proc_data.pid != w->pid || attached_proc_data.tid != w->tid)
+    if (proc.pid != w->pid || proc.tid != w->tid)
         return VMI_EVENT_RESPONSE_NONE;
 
     if (w->rsp && info->regs->rsp <= w->rsp)
@@ -1032,22 +1026,19 @@ static event_response_t createfile_ret_cb(drakvuf_t drakvuf, drakvuf_trap_info_t
         auto filename = get_file_name(w->f, drakvuf, vmi_lg.vmi, info, handle, nullptr, nullptr);
         if (filename.empty()) filename = "<UNKNOWN>";
 
-        w->f->files[{attached_proc_data.pid, handle}] = {filename, FILEEXTR_DELETE};
+        w->f->files[{proc.pid, handle}] = {filename, FILEEXTR_DELETE};
     }
+
+    delete w;
+    drakvuf_remove_trap(drakvuf, info->trap, (drakvuf_trap_free_t)g_free);
 
     return VMI_EVENT_RESPONSE_NONE;
 }
 
 static void createfile_cb_impl(drakvuf_t drakvuf, drakvuf_trap_info_t* info, addr_t handle)
 {
-    addr_t attached_proc = drakvuf_get_current_attached_process(drakvuf, info);
-    proc_data_t attached_proc_data;
-    if (!attached_proc ||
-        !drakvuf_get_process_data(drakvuf, attached_proc,
-                                  &attached_proc_data) ||
-        !drakvuf_get_current_thread_id(drakvuf, info,
-                                       &attached_proc_data.tid) ||
-        !attached_proc_data.pid || !attached_proc_data.tid)
+    attached_proc_data_t proc(drakvuf, info);
+    if (!proc.pid)
     {
         PRINT_DEBUG("[FILEDELETE2] [PID:%d] [TID:%d] Error: Failed to get "
                     "attached process\n",
@@ -1060,8 +1051,8 @@ static void createfile_cb_impl(drakvuf_t drakvuf, drakvuf_trap_info_t* info, add
     vmi_read_addr_va(vmi_lg.vmi, info->regs->rsp, 0, &ret_addr);
 
     auto w = new createfile_ret_info;
-    w->pid = attached_proc_data.pid;
-    w->tid = attached_proc_data.tid;
+    w->pid = proc.pid;
+    w->tid = proc.tid;
     w->rsp = info->regs->rsp;
     w->handle = handle;
     w->f = (filedelete*)info->trap->data;
@@ -1080,6 +1071,7 @@ static void createfile_cb_impl(drakvuf_t drakvuf, drakvuf_trap_info_t* info, add
     {
         printf("Failed to trap return at 0x%lx\n", ret_addr);
         delete w;
+        g_free(trap);
     }
 }
 
@@ -1121,14 +1113,8 @@ static event_response_t createfile_cb(drakvuf_t drakvuf, drakvuf_trap_info_t* in
  */
 static event_response_t setinformation_cb(drakvuf_t drakvuf, drakvuf_trap_info_t* info)
 {
-    addr_t attached_proc = drakvuf_get_current_attached_process(drakvuf, info);
-    proc_data_t attached_proc_data;
-    if (!attached_proc ||
-        !drakvuf_get_process_data(drakvuf, attached_proc,
-                                  &attached_proc_data) ||
-        !drakvuf_get_current_thread_id(drakvuf, info,
-                                       &attached_proc_data.tid) ||
-        !attached_proc_data.pid || !attached_proc_data.tid)
+    attached_proc_data_t proc(drakvuf, info);
+    if (!proc.pid)
     {
         PRINT_DEBUG("[FILEDELETE2] [PID:%d] [TID:%d] Error: Failed to get "
                     "attached process\n",
@@ -1161,7 +1147,7 @@ static event_response_t setinformation_cb(drakvuf_t drakvuf, drakvuf_trap_info_t
             auto filename = get_file_name(f, drakvuf, vmi, info, handle, nullptr, nullptr);
             if (filename.empty()) filename = "<UNKNOWN>";
 
-            f->files[{attached_proc_data.pid, handle}] = {filename, FILEEXTR_DELETE};
+            f->files[{proc.pid, handle}] = {filename, FILEEXTR_DELETE};
         }
     }
 
@@ -1172,14 +1158,8 @@ done:
 
 static event_response_t writefile_cb(drakvuf_t drakvuf, drakvuf_trap_info_t* info)
 {
-    addr_t attached_proc = drakvuf_get_current_attached_process(drakvuf, info);
-    proc_data_t attached_proc_data;
-    if (!attached_proc ||
-        !drakvuf_get_process_data(drakvuf, attached_proc,
-                                  &attached_proc_data) ||
-        !drakvuf_get_current_thread_id(drakvuf, info,
-                                       &attached_proc_data.tid) ||
-        !attached_proc_data.pid || !attached_proc_data.tid)
+    attached_proc_data_t proc(drakvuf, info);
+    if (!proc.pid)
     {
         PRINT_DEBUG("[FILEDELETE2] [PID:%d] [TID:%d] Error: Failed to get "
                     "attached process\n",
@@ -1195,7 +1175,7 @@ static event_response_t writefile_cb(drakvuf_t drakvuf, drakvuf_trap_info_t* inf
     auto filename = get_file_name(f, drakvuf, vmi, info, handle, nullptr, nullptr);
     if (filename.empty()) filename = "<UNKNOWN>";
 
-    f->files[{attached_proc_data.pid, handle}] = {filename, FILEEXTR_WRITE};
+    f->files[{proc.pid, handle}] = {filename, FILEEXTR_WRITE};
 
     drakvuf_release_vmi(drakvuf);
     return 0;
@@ -1210,14 +1190,8 @@ static event_response_t writefile_cb(drakvuf_t drakvuf, drakvuf_trap_info_t* inf
  */
 static event_response_t close_cb(drakvuf_t drakvuf, drakvuf_trap_info_t* info)
 {
-    addr_t attached_proc = drakvuf_get_current_attached_process(drakvuf, info);
-    proc_data_t attached_proc_data;
-    if (!attached_proc ||
-        !drakvuf_get_process_data(drakvuf, attached_proc,
-                                  &attached_proc_data) ||
-        !drakvuf_get_current_thread_id(drakvuf, info,
-                                       &attached_proc_data.tid) ||
-        !attached_proc_data.pid || !attached_proc_data.tid)
+    attached_proc_data_t proc(drakvuf, info);
+    if (!proc.pid)
     {
         PRINT_DEBUG("[FILEDELETE2] [PID:%d] [TID:%d] Error: Failed to get "
                     "attached process\n",
@@ -1228,7 +1202,7 @@ static event_response_t close_cb(drakvuf_t drakvuf, drakvuf_trap_info_t* info)
     filedelete* f = (filedelete*)info->trap->data;
     vmi_instance_t vmi = drakvuf_lock_and_get_vmi(drakvuf);
     addr_t handle = drakvuf_get_function_argument(drakvuf, info, 1);
-    auto reason = f->files[{attached_proc_data.pid, handle}].second;
+    auto reason = f->files[{proc.pid, handle}].second;
 
     event_response_t response = 0;
     if (f->use_injector)
@@ -1239,7 +1213,7 @@ static event_response_t close_cb(drakvuf_t drakvuf, drakvuf_trap_info_t* info)
         auto filename1 = get_file_name(f, drakvuf, vmi, info, handle, nullptr, nullptr);
         if (filename1.empty()) filename1 = "<UNKNOWN>";
 
-        auto filename = f->files[{attached_proc_data.pid, handle}].first;
+        auto filename = f->files[{proc.pid, handle}].first;
         if (filename.empty())
             goto done;
 
@@ -1247,7 +1221,7 @@ static event_response_t close_cb(drakvuf_t drakvuf, drakvuf_trap_info_t* info)
             goto done;
     }
 
-    if (f->files.erase({attached_proc_data.pid, handle}) > 0)
+    if (f->files.erase({proc.pid, handle}) > 0)
     {
         // We detect the fact of closing of the previously modified file.
         grab_file_by_handle(f, drakvuf, vmi, info, handle, reason);
@@ -1260,14 +1234,8 @@ done:
 
 static event_response_t createsection_cb(drakvuf_t drakvuf, drakvuf_trap_info_t* info)
 {
-    addr_t attached_proc = drakvuf_get_current_attached_process(drakvuf, info);
-    proc_data_t attached_proc_data;
-    if (!attached_proc ||
-        !drakvuf_get_process_data(drakvuf, attached_proc,
-                                  &attached_proc_data) ||
-        !drakvuf_get_current_thread_id(drakvuf, info,
-                                       &attached_proc_data.tid) ||
-        !attached_proc_data.pid || !attached_proc_data.tid)
+    attached_proc_data_t proc(drakvuf, info);
+    if (!proc.pid)
     {
         PRINT_DEBUG("[FILEDELETE2] [PID:%d] [TID:%d] Error: Failed to get "
                     "attached process\n",
@@ -1293,7 +1261,7 @@ static event_response_t createsection_cb(drakvuf_t drakvuf, drakvuf_trap_info_t*
     filename = get_file_name(f, drakvuf, vmi, info, handle, nullptr, nullptr);
     if (filename.empty()) filename = "<UNKNOWN>";
 
-    f->files[{attached_proc_data.pid, handle}] = {filename, FILEEXTR_WRITE};
+    f->files[{proc.pid, handle}] = {filename, FILEEXTR_WRITE};
 
 done:
     drakvuf_release_vmi(drakvuf);
